DoctorController was added

+++ appointment_system_Api/Model/Doctor.cs
@@ -7,13 +7,13 @@ namespace appointment_system_Api.Model
     public class Doctor
     {
         [Key]
-        public int NumEmpleado { get; set; }
+        public long NumEmpleado { get; set; }
         public string Name { get; set; }
         public string Last_Name { get; set; }
         public string Email { get; set; }
-        public int Phone { get; set; }
+        public long Phone { get; set; }
         public string Hospital_Name { get; set; }
-        public int NumeroCita { get; set; }
+        public long NumeroCita { get; set; }
 
     }
 }
